backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
boto3>=1.35.99
strands-agents>=1.0.0
strands-agents-tools>=0.2.19
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
httpx>=0.28.0
pytest>=8.3.4
pytest-asyncio>=0.25.2
